added comparables and to strings and stuffs

+++ AimsProject/src/hust/dsai/aims/order/Order.java
@@ -8,14 +8,14 @@ public class Order {
   int ord_id;
   public boolean status;
   float shippingCost;
-  public float totalCost;
+  public double totalCost;
 
   public Order(CustomerUser user, int id) {
     this.user = user;
     this.ord_id = id;
     this.status = false;
     this.shippingCost = AimsSystem.calculateShippingCost(this);
-    totalCost = user.cart.total * 1.1f + shippingCost;
+    totalCost = user.cart.total * 1.1 + shippingCost;
   }
 
   public void printOrder() {
