+++ AimsProject/src/hust/dsai/aims/transaction/Transaction.java
@@ -6,10 +6,10 @@ public class Transaction {
   String payment_info;
   int transaction_id;
   boolean transaction_status;
-  float total;
+  double total;
   Order order;
 
-  public Transaction(String payment_info, int transaction_id, float total, Order order) {
+  public Transaction(String payment_info, int transaction_id, double total, Order order) {
     this.payment_info = payment_info;
     this.transaction_id = transaction_id;
     this.transaction_status = false;
